move all parser into separate file in one folder
